Ingredients router

+++ src/recipe/recipe-service.js
@@ -17,25 +17,7 @@ const RecipeService = {
     return db('recipes')
       .select('*')
   },
-  createRecipe(db, object, id){
-    return db.into('recipes').insert(
-      {owner: id,
-      name: object.name,
-      category: object.category,
-      description: object.description,
-      ingredients: object.ingredients,
-      instructions: object.instructions,
-      time_to_make: object.time_to_make}
-    )
-  },
-  getMyRecipes(db, array){
-    let ret = []
-    array.map(id => {
-      ret.push(db('recipes').where({id}))
-    })
-    return ret
-  },
-  getUser(db, id){
+  getUser(db, id){ //should be in different service file
     return db('users')
       .where({id})
   },
@@ -45,15 +27,15 @@ const RecipeService = {
       .returning('*')
       .then(recipe => recipe[0])
   },
-  deleteRecipe(db, id) {
+  deleteRecipe(db, id, user_id) {
     return db('recipes')
       .delete()
-      .where({ id })
+      .where({ id, owner: user_id })
   },
-  editRecipe(db, id, newInfo) {
+  editRecipe(db, id, user_id, newInfo) {
     return db('recipes')
       .update(newInfo)
-      .where({ id })
+      .where({ id, owner: user_id })
   },
 }
